first commit, reinstate old inpur field comp. & make separate account input field folder
